Accept P3376

+++ 202411/Week10/P3376.cpp
@@ -35,7 +35,7 @@ bool bfs()
     }
     return false;
 }
-int dinic(int u, long long flow)
+long long dinic(int u, long long flow)
 {
     if (u == t)
         return flow;
